AppImage: bundle libjack.so.0 again
This effectively reverts 3c9b8293.

libjack.so.0 is a dependency also for different modules like portaudio
so if we do not want to depend on instance in the system, we must ship
our version.

+++ data/scripts/Linux-AppImage/create-appimage.sh
@@ -78,6 +78,9 @@ while read -r x; do
         EXCLUDE_LIST="$EXCLUDE_LIST $NAME"
 done < excludelist
 for n in $EXCLUDE_LIST; do
+        if [ "$n" = libjack.so.0 ]; then # JACK is currently handled in AppRun
+                continue
+        fi
         if [ -f "$APPPREFIX/lib/$n" ]; then
                 rm "$APPPREFIX/lib/$n"
         fi
@@ -95,7 +98,7 @@ if [ -f "$(echo $APPPREFIX/lib/libva.so.* | cut -d\  -f 1)" ]; then
         done
 fi
 
-cp "$srcdir/data/scripts/Linux-AppImage/AppRun" "$srcdir/data/ultragrid.png" $APPDIR
+cp -r "$srcdir/data/scripts/Linux-AppImage/AppRun" "$srcdir/data/scripts/Linux-AppImage/scripts" "$srcdir/data/ultragrid.png" $APPDIR
 cp "$srcdir/data/uv-qt.desktop" $APPDIR/cz.cesnet.ultragrid.desktop
 APPIMAGEUPDATETOOL=$(command -v appimageupdatetool-x86_64.AppImage || true)
 if [ -n "$APPIMAGEUPDATETOOL" ]; then
